requirements.txt
PyQt6>=6.5,<7
pypdf>=4.0.0,<6
pymupdf>=1.23.8,<2
pdf2image>=1.17.0,<2
Pillow>=10.0.0,<12
psutil>=5.9.0,<6
numpy>=1.24.0,<3
cryptography>=3.4